Move ChangelogEntry model to new changelog app
This makes the entire behaviour of the model reusable for other page types with a changelog
